# slightly faster ARM kernel