Before

|                              Method |                    FileName |        Mean |     Error |    StdDev | kiB/input | Speed (GB/s) |
|------------------------------------ |---------------------------- |------------:|----------:|----------:|---------- |------------- |
|      SimdBase64DecodingRealDataUTF8 | data/dns/swedenzonebase.txt | 18,531.0 us | 233.38 us | 239.67 us |       .35 |         1.89 |
|      SimdBase64DecodingRealDataUTF8 |                 data/email/ |    226.2 us |   8.19 us |   9.43 us |    131.77 |         8.74 |


After

|                              Method |                    FileName |        Mean |     Error |    StdDev | kiB/input | Speed (GB/s) |
|------------------------------------ |---------------------------- |------------:|----------:|----------:|---------- |------------- |
|      SimdBase64DecodingRealDataUTF8 | data/dns/swedenzonebase.txt | 18,414.7 us | 177.56 us | 189.98 us |       .35 |         1.91 |
|      SimdBase64DecodingRealDataUTF8 |                 data/email/ |    200.5 us |   1.51 us |   1.61 us |    131.77 |         9.86 |